Added PBORenderContextProvider.cs to give the option to render using Pixel Buffer Objects (Still some issues with it). Removed some GL calls from OpenGLControl and SceneView, that in my opinion should be left by the user, because it might cause confusion (Like OpenGL.GL_DEPTH_TEST). Note that this might require for the samples to be double checked! Added missing constants and function calls to OpenGLExtensionsJOG.cs for review. If approved then they can be added to OpenGL.cs or OpenGLExtensions.cs. Upgraded and added classes to SceneGraph.JOG. This includes a less restrictive and - in my opinion - more OGL natural ShaderProgram.cs, Color.cs,...
Added samples to share my experience with OpenGL for the last 2 months. It should mostly follow the OGL 3.3+ conventions, except for the Gradient Backgrounds, who still use deprecated functionalities. The samples which are currently working include "Translucency", "Particle Systems", "Use of materials", "Lines".
Note that for the examples, it is required to install GlmNET, which can be easely retrieved from Nuget in Visual Studio.